13 button 1 pin
With the help of this circuit you are able to attach 13 button or even more by utilizing only one pin of arduino which is very helpful when you have to use a keypad module in your project. But this project has one disadvantage as well and that is you can't click more than one button at a time. Also, this is the prototype and testing is still going on. Checkout the code for this project from arduino projects repository.